backend/requirements.txt
PyYAML
tqdm
uvicorn
fastapi
pydantic
numpy